requirements.txt
asgiref==3.8.1
Django==4.2.18
django-filter==24.3
djangorestframework==3.15.2
drf-yasg==1.21.8
importlib_metadata==8.5.0
inflection==0.5.1
Markdown==3.7
numpy==2.0.2
packaging==24.2
pandas==2.2.3
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
zipp==3.21.0
